requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.1
scikit-learn==1.3.0
scipy==1.11.1
openai==0.28.1
python-dotenv==1.0.0
sqlparse==0.4.4
streamlit-aggrid==0.3.4
streamlit-option-menu==0.3.2
wordcloud==1.9.2
networkx==3.1
pyvis==0.3.2
statsmodels==0.14.0